# Conflicts:
#	android/src/pl/space_marine/game/Animator.java
#	android/src/pl/space_marine/game/Renderer.java

+++ core/src/pl/space_marine/game/impediments/obstacles/Meteor.java
@@ -6,7 +6,7 @@ import pl.space_marine.game.impediments.Obstacle;
 public class Meteor extends Obstacle {
     private float size;
 
-    public Meteor(Image image, float collisionDamage, float direction, int speed, int x, int y) {
+    public Meteor(Image image, float collisionDamage, int direction, int speed, int x, int y) {
         super(image, null, collisionDamage, direction, speed, x, y);
     }
 }
