+++ core/src/pl/space_marine/game/impediments/obstacles/Satelite.java
@@ -4,7 +4,7 @@ import pl.space_marine.game.assets.Image;
 import pl.space_marine.game.impediments.Obstacle;
 
 public class Satelite extends Obstacle {
-    public Satelite(Image image, float collisionDamage, float direction, int speed, int x, int y) {
+    public Satelite(Image image, float collisionDamage, int direction, int speed, int x, int y) {
         super(image, null, collisionDamage, direction, speed, x, y);
     }
 }
